Amend types and improve way in which points are set

+++ src/hexagon.hpp
@@ -10,8 +10,8 @@
 
 namespace hexago {
 
-    typedef uint16_t hexagon_size_t;
-    typedef uint8_t hexagon_decay_t;
+    typedef int32_t hexagon_size_t;
+    typedef uint16_t hexagon_decay_t;
     typedef double seconds_alive_t;
 
     typedef struct hexagon_points_t {
